+++ server/src/helpers.ts
@@ -1296,6 +1296,8 @@ export const standardTypeIssueIssuesToMessage = (
 					return `Node address is missing`;
 				case StandardTypeIssue.BINDING_ON_BUS_NODE:
 					return `Binding should be used on bus type${issue.templateStrings.length ? 's' : ''}: ${issue.templateStrings.join(' or ')}`;
+				case StandardTypeIssue.EMPTY_RANGE_MISMATCH_ADDRESS_SIZE_CELLS:
+					return `Empty ranges property should only be used when address size cells is identical to the parent node`;
 			}
 		})
 		.join(' or ')